[refactor] separate drawer sub component
