+++ components/bookNote/drawer/PreNoteQa.tsx
@@ -1,15 +1,41 @@
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 
-import { QaPair } from "./DrawerWrapper";
+export interface QaPair {
+  question: string;
+  answer: string[];
+}
 
 interface DrawerPreProps {
-  qaPair: QaPair;
   drawerIdx: number;
 }
 
 export default function DrawerPre(props: DrawerPreProps) {
-  const { qaPair, drawerIdx } = props;
+  const { drawerIdx } = props;
+
+  const qaPair: QaPair = { question: "", answer: [""] };
+
+  switch (drawerIdx) {
+    case 1:
+      qaPair.question = "1. 이 책에 어떤 기대를 하고 계신가요?";
+      qaPair.answer = [
+        "상황에 따라 변하는 '동기'를 한 곳에 잡아 두고 싶고,\n앞으로 진행될 모든 업무에 대해 내가 이 일을 왜 하는지 명확하게 할 수 있는 힌트를 얻을 수 있다고 기대하고 있다.",
+      ];
+      break;
+    case 2:
+      qaPair.question = "2. 이 책의 핵심 메시지는 무엇일까요?";
+      qaPair.answer = [
+        "이 책의 핵심은 열정적이고 행복한 삶의 근원이 되는 '왜' 찾는 법을 개인, 조직에 맞게 탐색, 발견, 유지, 공유 방법들에 소개한다.\n나는 개인 수준에서 '왜'를 찾고, 유지할 수 있는 방법이 궁금하다.",
+      ];
+      break;
+    default:
+      qaPair.question = "3. 궁금한 내용부터 우선순위 질문 리스트를 만들어보세요";
+      qaPair.answer = [
+        "왜 why가 중요하다고 주장하는 것일까?",
+        "나 자신의 why를 발견하는 방법은 무엇일까?",
+        "Why를 항상 살아있게 유지하려면 어떻게 해야할까?",
+      ];
+  }
 
   return (
     <>
